+++ packages/ui/src/components/text-animations/index.js
@@ -1,3 +1,4 @@
 export { default as TypingReveal } from './typing-reveal/typing-reveal';
 export { default as LetterSpacingReveal } from './letter-spacing-reveal/letter-spacing-reveal';
 export { default as BouncingText } from './bouncing-text/bouncing-text';
+export { default as ScrambleReveal } from './scramble-reveal/scramble-reveal';